+++ src/main/java/com/apap/tugas1/repository/JabatanPegawaiDb.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.apap.tugas1.model.JabatanPegawaiModel;
 
-public interface JabatanPegawaiDb extends JpaRepository<JabatanPegawaiModel, BigInteger> {
+public interface JabatanPegawaiDb extends JpaRepository<JabatanPegawaiModel, Long> {
 
 }
